ref: Reorganize Day 1 files
