+++ lib/puzzles/2023/day01/input.txt
